Frontend schema v2 (#218)
* Hatch 338/number type client functions (#208)

* HATCH-338: Client query & mutate functions for number & integer types

* update tests

* v2 strict types

* typing for number attributes

* update tests, type issues

* fix tests, depcheck, typechecks, eslint

* update example/react-rest to use v2 schema

* config cleanup

* fix example tsconfig

* update package-lock

* update package-lock

* comment out e2e tests

* update package-locks for example repos

* update package-lock

* package-lock updates

* Hatch 348/v2 datetime frontend (#231)

* HATCH-348: frontend support for datetime v2

* merge in feat branch

* merge feat

* HATCH-348: v2 datetime + v2 update/create/hooks

* deleteOne & useDeleteOne

* component typing

* fix types & tests

* fixes for v2 demo deployment

* revert uuidv4 changes

* filter fixes for v2

* remove console logs

* version bumps for fe v2 schema: 0.1.0

* update readme v2 for frontend

* a little bit of typing

* wip

* wip type refactor

* refactor types for allowNullInfer

* add allowNulInfer for fe typing

* add allowNulInfer for fe typing

* resolve conflicts & sync with main

* merge main & resolve conflicts

* update frontend example to use v2

* update filter and displays (#288)

* Draft: hatch-350: relationships v2 frontend (#286)

* hatch-350: make belongsTo work in v2 frontend

* adding back relationships in rest-client package

* update comments

* hatch-350: relationships

* hatch-350: relationships

* v2 relationships

* fix tests, don't show many relationships in table by default

* publish fe packages

* fix for filtering on related records, publish

* fix tests

* fix namespaces (#296)

* bump versions (#298)

* hatch-419: UUID v2 frontend (#299)

* hatch-419: update uuid in core to track whether field or reference id

* update tests

* add client functions

* repurpose  to  control

* always set control.hidden=true when dealing with relationship

* hatch-419: uuid v2 frontend

* add types for QueryList & QueryOne : Include

* bump versions

* hatch-487: re-enable error case for namespaced fields param (#300)

* hatch-348: fix datetime filter missing value input (#302)

* hatch-348: fix datetime filter missing value input

* fix types

* version bump

* version bump

* update readme with latest v2 frontend changes

* merge main into feature (#303)

* Hatch 349/enums packages (#305)

* uncomment tests

* add to typing

* add TRequired generic to enum control

* add todo comments

* update generic type

* update test file

* core changes for enum values types

* version bump

* update rcj

* bumps

* bumps for real

* Version bumps

---------


* playwright tweaks

* fix: empty field for custom column (#308)

* fix: empty field for custom column

* bump versions

* bump versions

* hatch-435/remove theme provider from muiprovider (#310)

* hatch-382: number filtering (#315)

* hatch-382: number filtering

* version bumps

* fix: render props typing (#320)

* fix: render props typing

* add comment

* export CollectionState in react package

* version bumps

* fix: catch error on setClientPropertyValueFromResponse and log with m… (#324)

* fix: catch error on setClientPropertyValueFromResponse and log with more detail

* remove only

* version bumps

* update packagelock

* fix tests

* remove old hatchify-app code

* version bumps

---------

+++ packages/rest-client/src/services/promise/deleteOne/deleteOne.ts
@@ -1,15 +1,24 @@
-import type { Schemas, Source } from "../../types"
+import type { PartialSchema } from "@hatchifyjs/core"
+import type { FinalSchemas, GetSchemaNames, RestClient } from "../../types"
 import { notifySubscribers } from "../../store"
+import { SchemaNameNotStringError, schemaNameIsString } from "../../utils"
 
 /**
  * Deletes a resource in the data source and notifies subscribers.
  */
-export const deleteOne = async (
-  dataSource: Source,
-  allSchemas: Schemas,
-  schemaName: string,
+export const deleteOne = async <
+  const TSchemas extends Record<string, PartialSchema>,
+  const TSchemaName extends GetSchemaNames<TSchemas>,
+>(
+  dataSource: RestClient<TSchemas, TSchemaName>,
+  allSchemas: FinalSchemas,
+  schemaName: TSchemaName,
   id: string,
 ): Promise<void> => {
+  if (!schemaNameIsString(schemaName)) {
+    throw new SchemaNameNotStringError(schemaName)
+  }
+
   await dataSource.deleteOne(allSchemas, schemaName, id)
 
   notifySubscribers()
